requirements.txt
torch==2.2.0
torch_geometric==2.4.0
torchaudio==2.2.0
torchvision==0.17.0
biotite==0.37.0
fair-esm==2.0.1